# docs/swagger.yaml
basePath: /api
definitions:
...
  pannypal_internal_service_transaction_dto.PeriodInfo:
    properties:
      end_date:
        type: string
      month:
        type: integer
      start_date:
        type: string
      year:
        type: integer
    type: object
...
  dto.CategoryListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
    type: object
  dto.CategoryResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.CategorySummary:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      count:
        type: integer
      percentage:
        type: number
      total_amount:
        type: number
      type:
        $ref: '#/definitions/models.TransactionType'
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  dto.CreateTransactionRequest:
    properties:
      amount:
        type: number
      category_id:
        type: integer
      description:
        type: string
      phone_number:
        type: string
      type:
        enum:
        - INCOME
        - EXPENSE
        type: string
    required:
    - amount
    - type
    type: object
...
  dto.TransactionSummaryResponse:
    properties:
      balance:
        type: number
      category_summary:
        items:
          $ref: '#/definitions/dto.CategorySummary'
        type: array
      expense_count:
        type: integer
      income_count:
        type: integer
      period:
        $ref: '#/definitions/pannypal_internal_service_transaction_dto.PeriodInfo'
      total_expense:
        type: number
      total_income:
        type: number
      transaction_count:
        type: integer
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
  dto.UpdateTransactionRequest:
    properties:
      amount:
        type: number
      category_id:
        type: integer
      description:
        type: string
      type:
        enum:
        - INCOME
        - EXPENSE
        type: string
    type: object
...
  models.TransactionType:
    enum:
    - INCOME
    - EXPENSE
    type: string
    x-enum-varnames:
    - TypeIncome
    - TypeExpense
  types.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API lengkap untuk monitoring pengeluaran dan pemasukan cash flow
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cash Flow API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            $ref: '#/definitions/dto.CategoryListResponse'
      summary: Get all categories
      tags:
      - Category APIs
    post:
      consumes:
      - application/json
      description: Create a new category for transactions
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
      summary: Create new category
      tags:
      - Category APIs
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
      summary: Delete category
      tags:
      - Category APIs
    get:
      consumes:
      - application/json
      description: Get specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category retrieved successfully
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get category by ID
      tags:
      - Category APIs
    put:
      consumes:
      - application/json
      description: Update existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
      summary: Update category
      tags:
      - Category APIs
...
  /transactions:
    get:
      consumes:
      - application/json
      description: List all transactions with filters and pagination
      parameters:
      - description: User's phone number
        in: query
        name: phone_number
        type: string
      - description: Filter by type (INCOME/EXPENSE)
        in: query
        name: type
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Filter transactions from this date (ISO format)
        in: query
        name: start_date
        type: string
      - description: Filter transactions until this date (ISO format)
        in: query
        name: end_date
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transactions retrieved successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get all transactions
      tags:
      - Transaction APIs
    post:
      consumes:
      - application/json
      description: Create new income or expense transaction
      parameters:
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
      summary: Create new transaction
      tags:
      - Transaction APIs
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: User's phone number
        in: query
        name: phone_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
      summary: Delete transaction
      tags:
      - Transaction APIs
    get:
      consumes:
      - application/json
      description: Get specific transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: User's phone number
        in: query
        name: phone_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction retrieved successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get transaction by ID
      tags:
      - Transaction APIs
    put:
      consumes:
      - application/json
      description: Update existing transaction. All fields are optional
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: User's phone number
        in: query
        name: phone_number
        type: string
      - description: Updated transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction updated successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
      summary: Update transaction
      tags:
      - Transaction APIs
  /transactions/summary:
    get:
      consumes:
      - application/json
      description: Get summary of transactions with totals and category breakdown
      parameters:
      - description: User's phone number
        in: query
        name: phone_number
        type: string
      - description: Summary from this date
        in: query
        name: start_date
        type: string
      - description: Summary until this date
        in: query
        name: end_date
        type: string
      - description: Filter by specific month (1-12)
        in: query
        name: month
        type: integer
      - description: Filter by specific year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaction summary retrieved successfully
          schema:
            $ref: '#/definitions/dto.TransactionSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get transaction summary
      tags:
      - Transaction APIs
swagger: "2.0"